Refactoring/view and view model

+++ JUDA_iOS/JUDA/ViewModel/Auth/AuthViewModel.swift
@@ -6,25 +6,14 @@
 //
 
 import SwiftUI
+import PhotosUI
+
 import FirebaseCore
 import FirebaseFirestore
 import FirebaseAuth
 import GoogleSignIn
 import CryptoKit
 import AuthenticationServices
-
-// MARK: - Auth Provider Option
-enum AuthProviderOption: String {
-    case apple = "apple.com"
-    case google = "google.com"
-    case email = "password"
-}
-
-// MARK: - User Liked List (Posts / Drinks) Type
-enum UserLikedListType: String {
-    case posts = "likedPosts"
-    case drinks = "likedDrinks"
-}
 
 // MARK: - Auth ( 로그인 / 로그아웃 / 탈퇴 / 본인 계정 )
 @MainActor
@@ -56,13 +45,6 @@ final class AuthViewModel: ObservableObject {
     private let db = Firestore.firestore()
     private let postCollection = "posts"
     private let drinkCollection = "drinks"
-    
-    init() {
-        Task {
-            // 로그인이 되어있다면, 유저 정보 받아오기
-            if signInStatus { await getCurrentUser() }
-        }
-    }
     
     // 현재 유저 있는지 확인, uid 받기
     private func checkCurrentUserID() throws -> String {
@@ -131,81 +113,207 @@ final class AuthViewModel: ObservableObject {
             UIApplication.shared.open(settingsURL, options: [:], completionHandler: nil)
         }
     }
+    
+    // MyPage / Setting 에서 사용
+    // 유저 닉네임 수정 시, 2글자 이상 10글자 이하 && 기존 닉네임과 같은지 체크
+    func isChangeUserName(changeName: String) -> Bool {
+        guard let user = currentUser else {
+            return false
+        }
+        return changeName.count >= 2 && changeName.count <= 10 && user.userField.name != changeName
+    }
+    
+    // 유저 프로필 사진 변경 시, 사용되는 메서드
+    func updateImage(selectedPhotos: [PhotosPickerItem]) async throws -> UIImage {
+        guard let selectedPhoto = selectedPhotos.first else {
+            throw PhotosPickerImageLoadingError.noSelectedPhotos
+        }
+        if let data = try await selectedPhoto.loadTransferable(type: Data.self),
+           let uiImage = UIImage(data: data) {
+            return uiImage
+        } else {
+            throw PhotosPickerImageLoadingError.invalidImageData
+        }
+    }
+    
+    // 회원 가입 - ProfileSettingView 에서 모든 항목을 입력하고, '완료' 를 눌렀을 때 수행
+    func signInDoneButtonTapped(name: String, age: Int, profileImage: UIImage?, gender: String, notification: Bool) async {
+        let provider: String? = Auth.auth().currentUser?.providerData.first?.providerID
+        do {
+            // 애플의 경우 로그아웃이 된 상태라, 애플인 것을 google 이 아닌 경우로 체크
+            if provider != AuthProviderOption.google.rawValue {
+                // 재로그인
+                let signWithApple = SignInWithAppleHelper()
+                let appleIDCredential = try await signWithApple()
+                isLoading = true
+                await signInApple(appleIDCredential: appleIDCredential)
+                signInStatus = true
+            }
+            // 프로필 이미지 storage 저장
+            let url = await uploadProfileImageToStorage(image: profileImage)
+            // 유저 이름, 생일, 성별, 프로필, 알림 동의 등 forestore 에 저장
+            addUserDataToStore(
+                name: name,
+                age: age,
+                profileImageURL: url,
+                gender: gender,
+                notification: notification
+                )
+            // 유저 데이터 받기
+            await getCurrentUser()
+        } catch {
+            errorMessage = "회원가입에 문제가 발생했어요.\n다시 시도해주세요."
+            showError = true
+            print("error :: signInDoneButtonTapped", error.localizedDescription)
+        }
+    }
+    
+    // 회원 탈퇴 - authProviders 를 체크해서 apple or google 탈퇴 로직 수행
+    func deleteAccount() async -> Bool {
+        isLoading = true
+        guard let authProvider = currentUser?.userField.authProviders else {
+            isLoading = false
+            return false
+        }
+        var result: Bool
+        // 애플 유저일때, 탈퇴 로직
+        if authProvider == AuthProviderOption.apple.rawValue {
+            result = await deleteAppleAccount()
+        // 구글 유저일때, 탈퇴 로직
+        } else if authProvider == AuthProviderOption.google.rawValue {
+            result = await deleteGoogleAccount()
+        // ?? - 예외
+        } else {
+            result = false
+        }
+        isLoading = false
+        return result
+    }
 }
 
 // MARK: - User Fetch
 extension AuthViewModel {
+    // 데이터를 한번에 병렬로 받아오기 위해 사용하는 enum
+    private enum CurrentUserResult {
+        case userField(UserField?)
+        case posts([Post]?)
+        case likedPosts([Post]?)
+        case likedDrinks([Drink]?)
+        case notifications([UserNotification]?)
+    }
+    
     // 현재 CurrentUser : User 가져오기
     func getCurrentUser() async {
+        var userFieldResult: UserField?
+        var postsResult: [Post]?
+        var likedPostsResult: [Post]?
+        var likedDrinksResult: [Drink]?
+        var notificationsResult: [UserNotification]?
         do {
             let uid = try checkCurrentUserID()
-            
-            await withTaskGroup(of: Void.self) { taskGroup in
+            await withTaskGroup(of: CurrentUserResult.self) { taskGroup in
                 // 현재 유저 UserField 받아오기
-                taskGroup.addTask { await self.getCurrentUserField(uid: uid) }
+                taskGroup.addTask { .userField(await self.getCurrentUserField(uid: uid)) }
                 // 현재 유저 Posts 받아오기
-                taskGroup.addTask { await self.getCurrentUserPosts(uid: uid) }
+                taskGroup.addTask { .posts(await self.getCurrentUserPosts(uid: uid)) }
                 // 현재 유저 LikedPosts 받아오기
-                taskGroup.addTask { await self.getCurrentUserLikedPosts(uid: uid) }
+                taskGroup.addTask { .likedPosts(await self.getCurrentUserLikedPosts(uid: uid)) }
                 // 현재 유저 LikedDrinks 받아오기
-                taskGroup.addTask { await self.getCurrentUserLikedDrinks(uid: uid) }
+                taskGroup.addTask { .likedDrinks(await self.getCurrentUserLikedDrinks(uid: uid)) }
                 // 현재 유저 Notifications 받아오기
-                taskGroup.addTask { await self.getCurrentUserNotifications(uid: uid) }
+                taskGroup.addTask { .notifications(await self.getCurrentUserNotifications(uid: uid)) }
+                // taskGroup 종료 시, result 받아서 메서드 내부의 프로퍼티에 할당
+                for await result in taskGroup {
+                    switch result {
+                    case .userField(let userField):
+                        userFieldResult = userField
+                    case .posts(let posts):
+                        postsResult = posts
+                    case .likedPosts(let likedPosts):
+                        likedPostsResult = likedPosts
+                    case .likedDrinks(let likedDrinks):
+                        likedDrinksResult = likedDrinks
+                    case .notifications(let notifications):
+                        notificationsResult = notifications
+                    }
+                }
+                // 옵셔널 해제
+                guard let userField = userFieldResult,
+                      let posts = postsResult,
+                      let likedPosts = likedPostsResult,
+                      let likedDrinks = likedDrinksResult,
+                      let notifications = notificationsResult else { return }
+                // 유저에 데이터 값 할당
+                currentUser = User(userField: userField,
+                                   posts: posts,
+                                   likedPosts: likedPosts,
+                                   likedDrinks: likedDrinks,
+                                   notifications: notifications)
             }
         } catch {
             showError = true
             errorMessage = error.localizedDescription
-            print(errorMessage)
+            print("error :: getCurrentUser", error.localizedDescription)
         }
     }
     
     // 현재 유저 UserField 받아오기
-    private func getCurrentUserField(uid: String) async {
+    private func getCurrentUserField(uid: String) async -> UserField? {
         do {
-            currentUser?.userField = try await firebaseUserService.fetchUserFieldData(uid: uid)
+            let userField = try await firebaseUserService.fetchUserFieldData(uid: uid)
+            return userField
         } catch {
             errorMessage = error.localizedDescription
-            print(errorMessage)
+            print("error :: getCurrentUserField", error.localizedDescription)
+            return nil
         }
     }
     
     // 현재 유저 Posts 받아오기
-    func getCurrentUserPosts(uid: String) async {
+    func getCurrentUserPosts(uid: String) async -> [Post]? {
         do {
-            currentUser?.posts = try await firebaseUserService.fetchUserWrittenPosts(uid: uid)
+            let posts = try await firebaseUserService.fetchUserWrittenPosts(uid: uid)
+            return posts
         } catch {
             errorMessage = error.localizedDescription
-            print(errorMessage)
+            print("error :: getCurrentUserPosts", error.localizedDescription)
+            return nil
         }
     }
     
     // 현재 유저 LikedPosts 받아오기
-    private func getCurrentUserLikedPosts(uid: String) async {
+    private func getCurrentUserLikedPosts(uid: String) async -> [Post]? {
         do {
-            currentUser?.likedPosts = try await firebaseUserService.fetchUserLikedPosts(uid: uid)
+            let likedPosts = try await firebaseUserService.fetchUserLikedPosts(uid: uid)
+            return likedPosts
         } catch {
             errorMessage = error.localizedDescription
-            print(errorMessage)
+            print("error :: getCurrentUserLikedPosts", error.localizedDescription)
+            return nil
         }
     }
     
     // 현재 유저 LikedDrinks 받아오기
-    private func getCurrentUserLikedDrinks(uid: String) async {
+    private func getCurrentUserLikedDrinks(uid: String) async -> [Drink]? {
         do {
-            currentUser?.likedDrinks = try await firebaseUserService.fetchUserLikedDrink(uid: uid)
+            let likedDrinks = try await firebaseUserService.fetchUserLikedDrink(uid: uid)
+            return likedDrinks
         } catch {
             errorMessage = error.localizedDescription
-            print(errorMessage)
+            print("error :: getCurrentUserLikedDrinks", error.localizedDescription)
+            return nil
         }
     }
     
     // 현재 유저 Notifications 받아오기
-    private func getCurrentUserNotifications(uid: String) async {
+    private func getCurrentUserNotifications(uid: String) async -> [UserNotification]? {
         do {
-            currentUser?.notifications = try await firebaseUserService.fetchUserNotifications(uid: uid)
+            let notifications = try await firebaseUserService.fetchUserNotifications(uid: uid)
+            return notifications
         } catch {
             errorMessage = error.localizedDescription
-            print(errorMessage)
+            print("error :: getCurrentUserNotifications", error.localizedDescription)
+            return nil
         }
     }
 }
@@ -295,6 +403,17 @@ extension AuthViewModel {
         }
     }
     
+    // 유저 프로필 url 수정 ( url 이 없었는데, 생기는 경우 )
+    func updateUserProfileImageURL(url: URL?) async {
+        do {
+            guard let url = url else { return }
+            let uid = try checkCurrentUserID()
+            await firebaseAuthService.updateUserProfileImageURL(uid: uid, url: url)
+        } catch {
+            print("error :: updateUserProfileImageURL", error.localizedDescription)
+        }
+    }
+    
     // 유저 '알림 설정' 수정
     private func updateUserNotificationAllowed(systemSetting: UNNotificationSetting) async throws -> Bool {
         guard let user = currentUser else {
@@ -322,7 +441,7 @@ extension AuthViewModel {
 // MARK: - Upload / 데이터 저장
 extension AuthViewModel {
     // 유저 정보 저장
-    func addUserDataToStore(name: String, age: Int,
+    func addUserDataToStore(name: String, age: Int, profileImageURL: URL?,
                             gender: String, notification: Bool) {
         do {
             let uid = try checkCurrentUserID()
@@ -330,28 +449,33 @@ extension AuthViewModel {
                 userData: UserField(
                     name: name, age: age, gender: gender,
 					fcmToken: "", notificationAllowed: notification,
-                    profileImageURL: currentUser?.userField.profileImageURL,
+                    profileImageURL: profileImageURL ?? URL(string: ""),
                     authProviders: try getProviderOptionString()),
                 uid: uid)
         } catch {
-            print("error :: addUserDataToStore :", error.localizedDescription)
+            print("error :: addUserDataToStore", error.localizedDescription)
         }
     }
     
     // 유저 가입 시, 프로필 이미지 올리기 + 이미지 URL 저장
-    func uploadProfileImageToStorage(image: UIImage?) async {
+    func uploadProfileImageToStorage(image: UIImage?) async  -> URL? {
         do {
-            let uid = try checkCurrentUserID()
             guard let image = image else {
-                print("error :: uploadProfileImageToStorage : image X")
-                return
+                return nil
             }
-            try await fireStorageService.uploadImageToStorage(folder: .user, image: image, fileName: uid)
+            let uid = try checkCurrentUserID()
+            try await fireStorageService.uploadImageToStorage(folder: .user,
+                                                              image: image,
+                                                              fileName: uid)
             // 유저 프로필 받아오기
-            let url = try await fireStorageService.fetchImageURL(folder: .user, fileName: uid)
+            let url = try await fireStorageService.fetchImageURL(folder: .user,
+                                                                 fileName: uid)
+            //
             currentUser?.userField.profileImageURL = url
+            return url
         } catch {
-            print("error :: uploadProfileImageToStorage :", error.localizedDescription)
+            print("error :: uploadProfileImageToStorage", error.localizedDescription)
+            return nil
         }
     }
 }
@@ -404,7 +528,7 @@ extension AuthViewModel {
                 if isNewUser {
                     signOut()
                     self.isNewUser = true
-                    print("Fisrt ✨ - Apple Sign Up 🍎")
+                    print("First ✨ - Apple Sign Up 🍎")
                 } else {
                     print("Apple Sign In 🍎")
                     await getCurrentUser()
@@ -425,7 +549,7 @@ extension AuthViewModel {
     }
     
     // 회원가입 or 회원탈퇴 시, 재 로그인 - Apple
-    func reauthApple() async -> ASAuthorizationAppleIDCredential? {
+    private func reauthApple() async -> ASAuthorizationAppleIDCredential? {
         do {
             let signInWithAppleHelper = SignInWithAppleHelper()
             let appleIDCredential = try await signInWithAppleHelper()
@@ -438,20 +562,17 @@ extension AuthViewModel {
     }
     
     // 회원탈퇴 - Apple
-    func deleteAppleAccount() async -> Bool {
+    private func deleteAppleAccount() async -> Bool {
         do {
             guard try getProviderOptionString() == AuthProviderOption.apple.rawValue else { return false }
             try await firebaseAuthService.deleteAccountWithApple()
 			let uid = try checkCurrentUserID()
 			firebaseAuthService.deleteUserData(uid: uid)
             resetData()
-            isLoading = false
             return true
         } catch {
-            print("error :: \(error.localizedDescription)")
+            print("error :: deleteAppleAccount", error.localizedDescription)
             errorMessage = "회원탈퇴에 문제가 발생했어요.\n다시 시도해주세요."
-            showError = true
-            isLoading = false
             return false
         }
     }
@@ -480,14 +601,14 @@ extension AuthViewModel {
             // 신규 유저
             if isNewUser {
                 self.isNewUser = true
-                print("Fisrt ✨ - Google Sign Up 🤖")
+                print("First ✨ - Google Sign Up 🤖")
             } else {
                 print("Google Sign In 🤖")
                 await getCurrentUser()
                 self.signInStatus = true
             }
         } catch {
-            print("error :: \(error.localizedDescription)")
+            print("error :: signInWithGoogle", error.localizedDescription)
             errorMessage = "로그인에 문제가 발생했어요.\n다시 시도해주세요."
             showError = true
             resetData()
@@ -495,14 +616,18 @@ extension AuthViewModel {
     }
     
     // 회원탈퇴 - Google
-    func deleteGoogleAccount() {
-        // TODO: - 구글 탈퇴 추가
-		do {
-			let uid = try checkCurrentUserID()
-			firebaseAuthService.deleteUserData(uid: uid)
-		} catch {
-			print("error :: \(error.localizedDescription)")
-			errorMessage = "회원탈퇴에 문제가 발생했어요.\n다시 시도해주세요."
-		}
+    private func deleteGoogleAccount() async -> Bool {
+        do {
+            guard try getProviderOptionString() == AuthProviderOption.google.rawValue else { return false }
+            try await firebaseAuthService.deleteAccountWithGoogle()
+            let uid = try checkCurrentUserID()
+            firebaseAuthService.deleteUserData(uid: uid)
+            resetData()
+            return true
+        } catch {
+            print("error :: deleteGoogleAccount", error.localizedDescription)
+            errorMessage = "회원탈퇴에 문제가 발생했어요.\n다시 시도해주세요."
+            return false
+        }
     }
 }
